requirements.txt
asgiref==3.11.0
black==25.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
flake8==7.3.0
idna==3.11
inflection==0.5.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.1
pillow==12.0.0
platformdirs==4.5.1
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
